Refactored header.js  and moved Language Selector code to language-selector.js and added related constants
