+++ blocks/header/header.js
@@ -1,5 +1,7 @@
 import { getMetadata } from '../../scripts/aem.js';
 import { loadFragment } from '../fragment/fragment.js';
+import getLanguageSelector from './language-selector.js';
+
 import {
   fetchLanguageNavigation,
   fetchLanguagePlaceholders,
@@ -362,88 +364,6 @@ async function fetchingPlaceholdersData(placeholdersData) {
   makeImageClickableNSettingAltText();
   settingAltTextForSearchIcon();
 }
-const toggleExpandLanguageSelector = (e) => {
-  const toggleContainer = e.currentTarget;
-  const toggleContent = toggleContainer.querySelector('div.language-content');
-  if (!toggleContainer || !toggleContent) return;
-  const toggler = toggleContainer.querySelector('div.language-toggle');
-  const isExpanded = toggleContainer.classList.contains('nav-item-expanded-active');
-  if (e.type === 'click') {
-    toggleContainer.classList.toggle('nav-item-expanded-active');
-    toggler.setAttribute('aria-expanded', !isExpanded);
-    toggleContent.classList.toggle('nav-item-content-expanded');
-  } else if (e.type === 'mouseenter' && isDesktop.matches) {
-    toggleContainer.classList.add('nav-item-expanded-active');
-    toggler.setAttribute('aria-expanded', true);
-    toggleContent.classList.add('nav-item-content-expanded');
-  } else if (e.type === 'mouseleave') {
-    toggleContainer.classList.remove('nav-item-expanded-active');
-    toggler.setAttribute('aria-expanded', false);
-    toggleContent.classList.remove('nav-item-content-expanded');
-  }
-};
-
-// Get Language Selector Display Text based on screen size
-const getLanguageDisplayText = (placeholdersData, lang) => ((window.screen.width >= 768
-  && placeholdersData[lang])
-  ? placeholdersData[lang]
-  : lang);
-
-// Language toggle based on screen width
-const updateLanguageTextContent = (domElement, placeholdersData, lang) => {
-  domElement.textContent = getLanguageDisplayText(placeholdersData, lang);
-};
-
-const handleResize = (entries, placeholdersData, language) => {
-  entries.forEach((entry) => {
-    if (entry.target) {
-      const aElement = entry.target.querySelector('a');
-      if (aElement) {
-        updateLanguageTextContent(aElement, placeholdersData, language);
-      }
-    }
-  });
-};
-
-// Create a ResizeObserver instance with a callback
-const createResizeObserver = (placeholdersData, language) => new ResizeObserver((entries) => {
-  handleResize(entries, placeholdersData, language);
-});
-
-// create language selector domElement
-const fetchLanguageSelectorContent = (placeholdersData, metaLangContent, langCode) => {
-  const ulElement = ul();
-  if (metaLangContent && metaLangContent.split(',').length > 0) {
-    const langPairs = metaLangContent.split(',');
-    langPairs.forEach((pair) => {
-      const [language, url] = pair.split('|').map((part) => part.trim());
-      const liElement = li(
-        { class: `${langCode === language ? 'show-selected' : ''}` },
-        a(
-          { href: `${url}` },
-          getLanguageDisplayText(placeholdersData, language),
-        ),
-      );
-      ulElement.append(liElement);
-      // ResizeObserver for the liElement to show language code for mobile
-      const resizeObserver = createResizeObserver(placeholdersData, language);
-      resizeObserver.observe(liElement);
-    });
-  } else {
-    const liElement = li(
-      { class: 'show-selected' },
-      a(
-        { href: window.location.href },
-        getLanguageDisplayText(placeholdersData, langCode),
-      ),
-    );
-    ulElement.append(liElement);
-    // ResizeObserver for the liElement to show language code for mobile
-    const resizeObserver = createResizeObserver(placeholdersData, langCode);
-    resizeObserver.observe(liElement);
-  }
-  return ulElement;
-};
 
 function createListItemWithAnchor(item) {
   // Create the main list item
@@ -580,40 +500,6 @@ function createNavMenu(structuredNav, searchByCountryPlaceholder) {
   return menuOverlay;
 }
 
-const getLanguageSelector = (placeholdersData, lang) => {
-  const metaLangContent = getMetadata('alternate-url');
-  const languageMap = fetchLanguageSelectorContent(placeholdersData, metaLangContent, lang);
-
-  const languageSelectorContent = div(
-    { class: 'language-content' },
-    languageMap,
-  );
-
-  const languageToggle = div(
-    {
-      class: 'language-toggle',
-      'aria-expanded': 'false',
-    },
-  );
-
-  const langSelector = div(
-    {
-      class: 'language-container',
-      onclick: (e) => toggleExpandLanguageSelector(e),
-    },
-    languageToggle,
-  );
-  langSelector.append(languageSelectorContent);
-
-  // Show Language code in Mobile
-  const resizeObserver = new ResizeObserver(() => {
-    updateLanguageTextContent(languageToggle, placeholdersData, lang);
-  });
-
-  resizeObserver.observe(langSelector);
-
-  return langSelector;
-};
 export default async function decorate(block) {
   // load nav as fragment
   const navMeta = getMetadata('nav');
